# packages/contracts/openapi/core.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MustBeViral Core API",
    "version": "2026-07-12",
    "description": "Typed execution, programmatic access and platform REST contracts for the ViralGraph cleanroom Core Worker."
  },
  "servers": [{ "url": "/" }],
  "paths": {
...
    "/v1/canvases/{id}": {
      "get": {
        "operationId": "get_canvas_context",
        "summary": "Get canvas context",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": { "type": "string", "minLength": 8, "maxLength": 128 }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "minLength": 1, "maxLength": 200 }
          }
        ],
        "responses": {
          "200": {
            "description": "Operation completed.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GetCanvasContextSuccess" }
              }
            }
          },
          "default": {
            "description": "Operation failed with a typed API error.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorEnvelope" } }
            }
          }
        }
      }
    },
    "/v1/canvases/{id}/patches": {
      "post": {
        "operationId": "apply_canvas_patch",
        "summary": "Apply canvas patch",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": { "type": "string", "minLength": 8, "maxLength": 128 }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "minLength": 1, "maxLength": 200 }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": { "type": "string", "minLength": 1, "maxLength": 200 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ApplyCanvasPatchRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation completed.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApplyCanvasPatchSuccess" }
              }
            }
          },
          "default": {
            "description": "Operation failed with a typed API error.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorEnvelope" } }
            }
          }
        }
      }
    },
    "/v1/canvases/{id}/validate": {
      "post": {
        "operationId": "validate_graph",
        "summary": "Validate graph",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": { "type": "string", "minLength": 8, "maxLength": 128 }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "minLength": 1, "maxLength": 200 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ValidateGraphRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation completed.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidateGraphSuccess" }
              }
            }
          },
          "default": {
            "description": "Operation failed with a typed API error.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorEnvelope" } }
            }
          }
        }
      }
    },
    "/v1/canvases/{id}/quotes": {
      "post": {
        "operationId": "quote_run",
        "summary": "Quote run",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": { "type": "string", "minLength": 8, "maxLength": 128 }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "minLength": 1, "maxLength": 200 }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": { "type": "string", "minLength": 1, "maxLength": 200 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/QuoteRunRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Operation completed.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/QuoteRunSuccess" } }
            }
          },
          "default": {
            "description": "Operation failed with a typed API error.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorEnvelope" } }
            }
          }
        }
      }
    },
    "/v1/quotes/{id}/runs": {
      "post": {
        "operationId": "start_run",
        "summary": "Start run",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": { "type": "string", "minLength": 8, "maxLength": 128 }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "minLength": 1, "maxLength": 200 }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": { "type": "string", "minLength": 1, "maxLength": 200 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/StartRunRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "Operation completed.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/StartRunSuccess" } }
            }
          },
          "default": {
            "description": "Operation failed with a typed API error.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorEnvelope" } }
            }
          }
        }
      }
    },
...
    "/v1/artifacts/{id}": {
      "get": {
        "operationId": "get_artifact",
        "summary": "Get artifact",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": { "type": "string", "minLength": 8, "maxLength": 128 }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "minLength": 1, "maxLength": 200 }
          }
        ],
        "responses": {
          "200": {
            "description": "Operation completed.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GetArtifactSuccess" }
              }
            }
          },
          "default": {
            "description": "Operation failed with a typed API error.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorEnvelope" } }
            }
          }
        }
      }
    },
...
    "/v1/models/{id}": {
      "get": {
        "operationId": "explain_model",
        "summary": "Explain model",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": { "type": "string", "minLength": 8, "maxLength": 128 }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "minLength": 1, "maxLength": 200 }
          }
        ],
        "responses": {
          "200": {
            "description": "Operation completed.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExplainModelSuccess" }
              }
            }
          },
          "default": {
            "description": "Operation failed with a typed API error.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorEnvelope" } }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "clientCredentials": {
        "type": "oauth2",
        "flows": { "clientCredentials": { "tokenUrl": "/v1/oauth/token", "scopes": {} } }
      },
      "falSignature": { "type": "apiKey", "in": "header", "name": "X-Fal-Signature" }
    },
    "schemas": {
...
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "minLength": 1 },
              "message": { "type": "string", "minLength": 1 },
              "request_id": {
                "type": "string",
                "minLength": 8,
                "maxLength": 128,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$"
              },
              "retryable": { "type": "boolean" },
              "details": {
                "type": "object",
                "propertyNames": { "type": "string" },
                "additionalProperties": {}
              }
            },
            "required": ["code", "message", "request_id", "retryable"],
            "additionalProperties": false
          }
        },
        "required": ["error"],
        "additionalProperties": false
      },
...
      "GetCanvasContextSuccess": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "canvas": {
                "type": "object",
                "properties": {
                  "canvasId": { "type": "string", "minLength": 1, "maxLength": 200 },
                  "projectId": { "type": "string", "minLength": 1, "maxLength": 200 },
                  "headRevisionId": { "type": "string", "minLength": 1, "maxLength": 200 },
                  "graphSchemaVersion": {
                    "type": "integer",
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991
                  },
                  "graphSnapshot": {
                    "type": "object",
                    "properties": {
                      "nodes": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "string", "minLength": 1, "maxLength": 200 },
                            "kind": {
                              "type": "string",
                              "enum": [
                                "brief",
                                "brand_context",
                                "planner_text",
                                "image_generation",
                                "image_edit",
                                "video_generation",
                                "qa",
                                "output_export",
                                "group"
                              ]
                            },
                            "parameter_schema_version": {
                              "type": "integer",
                              "exclusiveMinimum": 0,
                              "maximum": 9007199254740991
                            },
                            "parameters": {
                              "type": "object",
                              "propertyNames": { "type": "string" },
                              "additionalProperties": { "$ref": "#/$defs/__schema0" }
                            }
                          },
                          "required": ["id", "kind", "parameter_schema_version", "parameters"],
                          "additionalProperties": false
                        }
                      },
                      "edges": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "string", "minLength": 1, "maxLength": 200 },
                            "kind": { "type": "string", "const": "dependency" },
                            "source_node_id": {
                              "type": "string",
                              "minLength": 1,
                              "maxLength": 200
                            },
                            "target_node_id": { "type": "string", "minLength": 1, "maxLength": 200 }
                          },
                          "required": ["id", "kind", "source_node_id", "target_node_id"],
                          "additionalProperties": false
                        }
                      }
                    },
                    "required": ["nodes", "edges"],
                    "additionalProperties": false
                  },
                  "canonicalHash": { "type": "string", "pattern": "^[a-f0-9]{64}$" }
                },
                "required": [
                  "canvasId",
                  "projectId",
                  "headRevisionId",
                  "graphSchemaVersion",
                  "graphSnapshot",
                  "canonicalHash"
                ],
                "additionalProperties": false
              }
            },
            "required": ["canvas"],
            "additionalProperties": false
          },
          "meta": {
            "type": "object",
            "properties": {
              "request_id": {
                "type": "string",
                "minLength": 8,
                "maxLength": 128,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$"
              }
            },
            "required": ["request_id"],
            "additionalProperties": false
          }
        },
        "required": ["data", "meta"],
        "additionalProperties": false,
        "$defs": {
          "__schema0": {
            "anyOf": [
              { "type": "string" },
              { "type": "number" },
              { "type": "boolean" },
              { "type": "null" },
              { "type": "array", "items": { "$ref": "#/$defs/__schema0" } },
              {
                "type": "object",
                "propertyNames": { "type": "string" },
                "additionalProperties": { "$ref": "#/$defs/__schema0" }
              }
            ]
          }
        }
      },
      "ApplyCanvasPatchSuccess": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "canvasId": { "type": "string", "minLength": 1, "maxLength": 200 },
              "revisionId": { "type": "string", "minLength": 1, "maxLength": 200 },
              "canonicalHash": { "type": "string", "pattern": "^[a-f0-9]{64}$" },
              "affectedDescendants": {
                "type": "array",
                "items": { "type": "string", "minLength": 1, "maxLength": 200 }
              }
            },
            "required": ["canvasId", "revisionId", "canonicalHash", "affectedDescendants"],
            "additionalProperties": false
          },
          "meta": {
            "type": "object",
            "properties": {
              "request_id": {
                "type": "string",
                "minLength": 8,
                "maxLength": 128,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$"
              }
            },
            "required": ["request_id"],
            "additionalProperties": false
          }
        },
        "required": ["data", "meta"],
        "additionalProperties": false
      },
      "ApplyCanvasPatchRequest": {
        "type": "object",
        "properties": {
          "expected_revision_id": { "type": "string", "minLength": 1, "maxLength": 200 },
          "reason": { "type": "string", "minLength": 1, "maxLength": 500 },
          "patch": {
            "type": "object",
            "properties": {
              "upsert_nodes": {
                "default": [],
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "minLength": 1, "maxLength": 200 },
                    "kind": {
                      "type": "string",
                      "enum": [
                        "brief",
                        "brand_context",
                        "planner_text",
                        "image_generation",
                        "image_edit",
                        "video_generation",
                        "qa",
                        "output_export",
                        "group"
                      ]
                    },
                    "parameter_schema_version": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    },
                    "parameters": {
                      "type": "object",
                      "propertyNames": { "type": "string" },
                      "additionalProperties": { "$ref": "#/$defs/__schema0" }
                    }
                  },
                  "required": ["id", "kind", "parameter_schema_version", "parameters"],
                  "additionalProperties": false
                }
              },
              "remove_node_ids": {
                "default": [],
                "type": "array",
                "items": { "type": "string", "minLength": 1, "maxLength": 200 }
              },
              "upsert_edges": {
                "default": [],
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "minLength": 1, "maxLength": 200 },
                    "kind": { "type": "string", "const": "dependency" },
                    "source_node_id": { "type": "string", "minLength": 1, "maxLength": 200 },
                    "target_node_id": { "type": "string", "minLength": 1, "maxLength": 200 }
                  },
                  "required": ["id", "kind", "source_node_id", "target_node_id"],
                  "additionalProperties": false
                }
              },
              "remove_edge_ids": {
                "default": [],
                "type": "array",
                "items": { "type": "string", "minLength": 1, "maxLength": 200 }
              }
            },
            "required": ["upsert_nodes", "remove_node_ids", "upsert_edges", "remove_edge_ids"],
            "additionalProperties": false
          }
        },
        "required": ["expected_revision_id", "reason", "patch"],
        "additionalProperties": false,
        "$defs": {
          "__schema0": {
            "anyOf": [
              { "type": "string" },
              { "type": "number" },
              { "type": "boolean" },
              { "type": "null" },
              { "type": "array", "items": { "$ref": "#/$defs/__schema0" } },
              {
                "type": "object",
                "propertyNames": { "type": "string" },
                "additionalProperties": { "$ref": "#/$defs/__schema0" }
              }
            ]
          }
        }
      },
...
      "ValidateGraphRequest": { "type": "object", "properties": {}, "additionalProperties": false },
...
      "QuoteRunRequest": {
        "type": "object",
        "properties": {
          "expected_revision_id": { "type": "string", "minLength": 1, "maxLength": 200 }
        },
        "required": ["expected_revision_id"],
        "additionalProperties": false
      },
...
      "StartRunRequest": {
        "type": "object",
        "properties": {
          "confirmed": { "type": "boolean", "const": true },
          "confirmation_token": { "type": "string", "minLength": 16, "maxLength": 500 }
        },
        "required": ["confirmed", "confirmation_token"],
        "additionalProperties": false
      },
...
    }
  }
}